modal/pyproject.toml
[project]
name = "robogen-modal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "modal>=1.2.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
    "networkx>=3.4.2",
    "matplotlib>=3.10.0",
    "scikit-learn>=1.8.0",
    "openai>=2.15.0",
]
